moves last remaining used funcs from split.rs to core.rs
